refactoring task7
