+++ ang/course/src/app/app-routing.module.ts
@@ -1,18 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { UsersComponent } from './display/users/users.component';
-import { AdduserComponent } from './display/adduser/adduser.component';
-import { EdituserComponent } from './display/edituser/edituser.component';
-import { LoginComponent } from './auth/components/login/login.component';
-import { RegisterComponent } from './auth/components/register/register.component';
 
 const routes: Routes = [
-  {path:'users', component: UsersComponent},
-  {path:'add', component: AdduserComponent},
-  {path:'login', component: LoginComponent},
-  {path:'register', component: RegisterComponent},
-  {path:'edit/:id', component: EdituserComponent},
-  {path:'', redirectTo: '/login', pathMatch: 'full'}
+  // {path:'users', component: UsersComponent},
+  // {path:'add', component: AdduserComponent},
+  // {path:'login', component: LoginComponent},
+  // {path:'register', component: RegisterComponent},
+  // {path:'edit/:id', component: EdituserComponent},
+  // {path:'', redirectTo: '/login', pathMatch: 'full'}
 ];
 
 @NgModule({
