+++ ang/course/src/app/app.component.html
@@ -1,4 +1,3 @@
-<app-header></app-header>
 
 <div class="app-container">
     <router-outlet></router-outlet>
