+++ ang/course/src/app/app.module.ts
@@ -4,28 +4,14 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HeaderComponent } from './display/header/header.component';
-import { UsersComponent } from './display/users/users.component';
-import { AdduserComponent } from './display/adduser/adduser.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {MatIconModule} from '@angular/material/icon'
 import {MatFormFieldModule} from '@angular/material/form-field';
-import { EdituserComponent } from './display/edituser/edituser.component';
-import { UserComponent } from './forms/user/user.component';
-import { AddressComponent } from './forms/address/address.component';
-import { UserlistComponent } from './display/userlist/userlist.component';
 import { AuthModule } from './auth/auth.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    HeaderComponent,
-    UsersComponent,
-    AdduserComponent,
-    EdituserComponent,
-    UserComponent,
-    AddressComponent,
-    UserlistComponent,
   ],
   imports: [
     BrowserModule,
