--- ang/course/src/app/display/adduser/adduser.component.html
@@ -1,1 +1,0 @@
-<app-user [newUser]="true"></app-user>
