--- ang/course/src/app/display/edituser/edituser.component.html
@@ -1,1 +1,0 @@
-<app-user [newUser]="false"></app-user>
