--- ang/course/src/app/display/users/users.component.css
@@ -1,78 +1,0 @@
-.container-box{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.user-card{
-    height: 200px;
-    width: 200px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-}
-
-button{
-    font-size: 12px;
-}
-
-.cards{
-    display: flex;
-    gap: 3em;
-}
-
-.active-name{
-    background-color: rgb(111, 158, 111);
-    border-radius: 5px;
-    border: 2px solid rgb(88, 126, 96);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-.inactive-name{
-    background-color: rgb(115, 117, 115);
-    border-radius: 5px;
-    border: 2px solid rgb(87, 88, 87);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-.active-only-button{
-    height: 30px;
-    width: 100px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 50px;
-    background-color: rgb(54, 54, 54);
-    border: 1px solid black;
-    border-radius: 3px;
-    color: #fff;
-    font-weight: 700;
-    cursor: pointer;
-}
-
-.toggle-btn{
-    height: 30px;
-    width: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: 5px 0;
-    background-color: rgb(54, 54, 54);
-    border: 1px solid black;
-    border-radius: 3px;
-    color: #fff;
-    font-weight: 700;
-    cursor: pointer;
-}
-
-mat-icon{
-    cursor: pointer;
-}
