--- ang/course/src/app/forms/user/user.component.html
@@ -1,112 +1,0 @@
-<h2>Add user</h2>
-<form [formGroup]="userform">
-    <div class="form-label">
-        <label>
-            First Name:
-        </label>
-        <input type="text" formControlName="firstName" class="input-text">
-        <div *ngIf="this.userform.get('firstName').errors?.['required']" class="red">
-            First Name is required.
-        </div>
-    </div>
-
-    <div class="form-label">
-        <label>
-            Last Name:
-        </label>
-        <input type="text" formControlName="lastName" class="input-text" required>
-        <div *ngIf="this.userform.get('lastName').errors?.['required']" class="red">
-            Last Name is required.
-        </div>
-    </div>
-    
-    <div class="form-label">
-        <label>
-            Age:
-        </label>
-        <input type="text" formControlName="age" class="input-text" >
-        <div *ngIf="this.userform.get('age').errors?.['required']" class="red">
-            Age is required.
-        </div>
-        <div *ngIf="this.userform.get('age').errors?.['min']" class="red">
-            Age has to be over 15.
-        </div>
-        <div *ngIf="this.userform.get('age').errors?.['max']" class="red">
-            Age has to be under 100.
-        </div>
-    </div>
-
-    <div class="form-label">
-        <label>
-            Company:
-        </label>
-        <input type="text" formControlName="company" class="input-text">
-        <div *ngIf="this.userform.get('company').errors?.['maxLength']" class="red">
-            Company has to be under 35 characters long.
-        </div>
-        <div *ngIf="this.userform.get('company').errors?.['required']" class="red">
-            Company is required.
-        </div>
-    </div>
-
-    <div class="form-label">
-        <label>
-            Department:
-        </label>
-        <select name="departments" id="department-select" formControlName="department">
-            <option value="frontend">Front-end</option>
-            <option value="backend">Back-end</option>
-        </select>
-    </div>
-
-    <div class="form-label">
-        <label>
-            E-mail address:
-        </label>
-        <input type="text" formControlName="email" class="input-text">
-        <div *ngIf="this.userform.get('email').errors?.['required']" class="red">
-            E-mail is required.
-        </div>
-        <div *ngIf="this.userform.get('email').errors?.['email']" class="red">
-            E-mail format not valid.
-        </div>
-        <div *ngIf="this.userform.get('email').errors?.['pattern']" class="red">
-            Only Gmail allowed.
-        </div>
-        <div *ngIf="this.userform.get('email').errors?.['doesEmailExist']" class="red">
-            E-mail already registered.
-        </div>
-    </div>
-
-    <app-address></app-address>
-
-    <div class="form-radio">
-        <fieldset>
-            <legend>Gender:</legend>
-            <div>
-                <input type="radio" id="male" name="gender" value="Male" formControlName="gender" checked>
-                <label for="male">Male</label>
-            </div>
-            <div>
-                <input type="radio" id="female" name="gender" value="Female" formControlName="gender">
-                <label for="female">Female</label>
-            </div>
-            
-        </fieldset>
-        <div *ngIf="this.userform.get('gender').errors?.['required']" class="red">
-            Gender is required.
-        </div>
-    </div>
-
-    <div class="form-radio">
-        <input type="checkbox" formControlName="active">
-        <label>
-            Active
-        </label>
-    </div>
-
-
-    <div *ngIf="newUser==true"><button type="submit" (click)='addUserToList()' [disabled]="!userform.valid" routerLink="/users">Add</button></div>
-    <div *ngIf="newUser==false"><button type="submit" (click)='updateUser()' [disabled]="!userform.valid" routerLink="/users">Update</button></div>
-    <div *ngIf="!userform.valid" class="red">Form is incomplete!</div>
-</form>
