+++ ang/course/src/app/shared/shared.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HeaderBarComponent } from './components/header-bar/header-bar.component';
+
+
+
+@NgModule({
+  declarations: [
+    HeaderBarComponent
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class SharedModule { }
