+++ ang/course/src/app/user/containers/add-user-shell/add-user-shell.component.html
@@ -1,0 +1,1 @@
+<app-user-form [newUser]="true"></app-user-form>
