+++ ang/course/src/app/user/containers/add-user-shell/add-user-shell.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-add-user-shell',
+  templateUrl: './add-user-shell.component.html',
+  styleUrls: ['./add-user-shell.component.css']
+})
+export class AddUserShellComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
