+++ ang/course/src/app/user/containers/edit-user-shell/edit-user-shell.component.html
@@ -1,0 +1,1 @@
+<app-user-form [newUser]="false"></app-user-form>
